:recycle: refactor: move task creation files to their own folder
